Magnetometer Homing Changes (V1)
I have implemented an addition to the (Circle) homing routine where the QB turret will spin in a circle after completing the laser homing. It tracks max and min X and Y values, effectively calibrating itself so that it can then print out the heading that it sees at any given time. This is accurate to at least 10 degrees from what I was testing probably more like 3 or 5.