Revert "Testing acceleration rate"
This reverts commit 11cc12117d7f0b931003b4d86eb3b866cf369798.

We now believe the acceleration rate was affected by the serial configuration for the sabertooth, and thus this change was not necessary.

By default, it was not using
"Map 1" exponential mapping, and was using no mapping at all. According to the configuration for 2.72, using PWM, Map 1 is used for the output pins to the motors.